permisos y gestion de vales

+++ app/Http/Controllers/NominasController.php
@@ -9,6 +9,7 @@ use App\Traits\DatosimpleTraits;
 use App\Models\Nominas_pagosenc;
 use App\Models\tarifasisr_det;
 use App\Models\Nominas_pagosdet;
+use App\Traits\SistemasTraits;
 use DB;
 use Maatwebsite\Excel\Facades\Excel;
 use App\Imports\BoTranImport;
@@ -20,6 +21,7 @@ class NominasController extends Controller
     
     use MenuTrait;
     use DatosimpleTraits;
+    use SistemasTraits;
 
     public function __construct()
     {
@@ -33,17 +35,25 @@ class NominasController extends Controller
     $varsubmenus =   $this->Traermenudet();
     $varnominas =  $this->obtenernominas();
     $varisrenc =  $this->obtenerisrenc();
-    
     $date = Carbon::now();
     $date = $date->format('Y-m-d');
-        return view('nominas.nomina',compact('varpantallas','varsubmenus','varnominas','varisrenc'));
+    $idusuario=auth()->user()->id;
+    $permisos = $this->obtener_permisosxusuario($idusuario);
+    return view('nominas.nomina',compact('varpantallas','varsubmenus','varnominas','varisrenc','permisos'));
+    
     }
+
+
+
+
 
     public function store(Request $request)
     {
+        $permisos = $this->forpermisos('alta_nominas');  
+        if($permisos=="alta_nominas")
+        { 
         $date = Carbon::now();
         $fecha = $date->format('Y-m-d');
-    
 
         //insertamos en la tabla pagosnominasencabezado
         $nominas_pagoenc = new Nominas_pagosenc();
@@ -54,34 +64,56 @@ class NominasController extends Controller
         $nominas_pagoenc->comentarios='comentario';
         $nominas_pagoenc->idtiponomina=$request->get('tipo_nomina');
         $nominas_pagoenc->save();
+        
         return redirect()->route('vernominas')->with("success","Nomina Creada Correctamente");
+        }
+        else{
+            return redirect()->route('vernominas')->with("Errorpermisos","Nomina Error");
+        }
+    
+
     }
 
     public function insertarnomina($id,$idpago_nomina){
-        $dias_nomina=0;
-        if($idpago_nomina==1){
-            $dias_nomina = 7;
-        }
-        if($idpago_nomina==2){
-            $dias_nomina = 15;
-        }
-        if($idpago_nomina==3){
-            $dias_nomina = 30;
-        }
 
-        //mandamos llamar el procedimiento de calculonomima
-        $pagosnom = DB::select('CALL insertar_nominapag_det(?,?)', [$id,$dias_nomina]);
-        $pagosnomina = DB::select('CALL generarcalculosnomina(?)', [$id]);
-        $Borrartablatemp =  DB::select('truncate table temptblnominas_pagodet');
-        //actualizamos el estado de tabla pagosnominaencabezado
-        $pagonomenc = Nominas_pagosenc::find($id);
-        $pagonomenc->estado_nomina = 'Edicion';
-        $pagonomenc->save();
-        return redirect()->route('vernominas')->with("successcalcular","Nomina Creada Correctamente");
+        $permisos = $this->forpermisos('calcular_nominas');  
+            if($permisos=="calcular_nominas")
+            {
+                $dias_nomina=0;
+                if($idpago_nomina==1){
+                    $dias_nomina = 7;
+                }
+                if($idpago_nomina==2){
+                    $dias_nomina = 15;
+                }
+                if($idpago_nomina==3){
+                    $dias_nomina = 30;
+                }
+        
+                //mandamos llamar el procedimiento de calculonomima
+                $pagosnom = DB::select('CALL insertar_nominapag_det(?,?)', [$id,$dias_nomina]);
+                $pagosnomina = DB::select('CALL generarcalculosnomina(?)', [$id]);
+                $Borrartablatemp =  DB::select('truncate table temptblnominas_pagodet');
+                //actualizamos el estado de tabla pagosnominaencabezado
+                $pagonomenc = Nominas_pagosenc::find($id);
+                $pagonomenc->estado_nomina = 'Edicion';
+                $pagonomenc->save();
+                return redirect()->route('vernominas')->with("successcalcular","Nomina Creada Correctamente");
+            }
+            else{
+                return redirect()->route('vernominas')->with("Errorpermisos","No se logro");  
+            }
+
+      
     }
 
 
     public function cerrarnomina($id,$idtipodias ){
+
+    $permisos = $this->forpermisos('cerrar_nominas'); 
+    if($permisos=="cerrar_nominas")
+    {
+
         $dias_nomina=0;
         if($idtipodias==1){
             $dias_nomina = 7;
@@ -102,11 +134,21 @@ class NominasController extends Controller
         if($pagonomenc->save()){
             return redirect()->route('vernominas')->with("success","¡Se guardaron los cambios correctamente!");
         }
-        return back()->with("warning","No se logro");
-    
+        else{
+            return back()->with("warning","No se logro");
+        }
+    }
+    else
+    {
+        return redirect()->route('vernominas')->with("Errorpermisos","No se logro");  
+    }
     }
 
-    public function calcularnomina($id,$idtipodias ){
+    public function calcularnomina($id,$idtipodias){
+
+        $permisos = $this->forpermisos('calcular_nominas'); 
+        if($permisos=="calcular_nominas")
+        {
         $dias_nomina=0;
         if($idtipodias==1){
             $dias_nomina = 7;
@@ -120,40 +162,74 @@ class NominasController extends Controller
         $pagosnom = DB::select('CALL generarcalculosnominafinal(?)', [$id]);
         $Borrartbl =  DB::select('delete from tblnominas_pagodet where idpagonomina = ? ', [$id]);
         $pagosnomina = DB::select('CALL generarcalculosnomina(?)', [$id]);
-
         return back()->with("success","¡Se guardaron los cambios correctamente!");
+        }
+
+        else{
+            return redirect()->route('vernominas')->with("Errorpermisos","No se logro");  
+        }
     
     }
 
     public function nominaeliminar($id){
+
+        $permisos = $this->forpermisos('eliminar_nominas'); 
+        if($permisos=="eliminar_nominas")
+        {
         // solo de enc
         $Borrartbl =  DB::select('delete from tblnominas_pagoenc where id = ? ', [$id]);
         return back()->with("success","¡Se guardaron los cambios correctamente!");
+        }
+        else{
+            return redirect()->route('vernominas')->with("Errorpermisos","No se logro");   
+        }
     }
 
     public function nominaeliminarTemp($id){
+        
+        $permisos = $this->forpermisos('eliminar_nominas'); 
+        if($permisos=="eliminar_nominas")
+        {
         // solo de enc det temdet
         $Borrartbl =  DB::select('delete from tblnominas_pagoenc where id = ? ', [$id]);
         $Borrartbl =  DB::select('delete from tblnominas_pagodet where idpagonomina = ? ', [$id]);
         $Borrartbl =  DB::select('delete from temptblnominas_pagodet where idpagonomina = ? ', [$id]);
         return back()->with("success","¡Se guardaron los cambios correctamente!");
     }
+ else{
+    return redirect()->route('vernominas')->with("Errorpermisos","No se logro");  
+     }
+    }
 
     public function nominaeliminarCalcu($id){
+
+        $permisos = $this->forpermisos('eliminar_nominas'); 
+            if($permisos=="eliminar_nominas")
+            {
         // solo de enc det
         $Borrartbl =  DB::select('delete from tblnominas_pagoenc where id = ? ', [$id]);
         $Borrartbl =  DB::select('delete from tblnominas_pagodet where idpagonomina = ? ', [$id]);
-        return back()->with("success","¡Se guardaron los cambios correctamente!");
+        return back()->with("success","¡Se guardaron los cambios correctamente!"); 
+        }
+        else{
+            return redirect()->route('vernominas')->with("Errorpermisos","No se logro");  
+        }
         
     }
 
     public function editarnomina($id, $idtiponomina){
+
+        $permisos = $this->forpermisos('actualizar_nominas'); 
+         if($permisos=="actualizar_nominas")
+        {
        $varnominas =  $this->obtenernominasporid($id);
        $varpantallas =  $this->Traermenuenc();
        $varsubmenus =   $this->Traermenudet();
-       $varisr =  $this->obtenerisrdet($idtiponomina);
-       $obtenersub =  $this->obtenersubsisdiordet($idtiponomina);
-       return view('nominas.editarnomina',compact('varpantallas','varsubmenus','varnominas','varisr','obtenersub'));
+       return view('nominas.editarnomina',compact('varpantallas','varsubmenus','varnominas'));
+       }
+       else{
+        return redirect()->route('vernominas')->with("Errorpermisos","No se logro");  
+       }
     }
 
     public function editarnomemp($id,$idemple){
@@ -166,6 +242,9 @@ class NominasController extends Controller
 
     public function actualizarNominaEmp(request $request)
     {
+        $permisos = $this->forpermisos('actualizar_nominas_empleado'); 
+            if($permisos=="actualizar_nominas_empleado")
+            {
            $date = Carbon::now();
            $fecha = $date->format('Y-m-d');
            $id = $request->get('idpadodet');
@@ -181,11 +260,20 @@ class NominasController extends Controller
             return back()->with("success","¡Se guardaron los cambios correctamente!");
         }
         return back()->with("warning","No se logro");
+            
+        }
+        else{
+            return redirect()->route('vernominas')->with("Errorpermisos","No se logro");  
+        }
     }
 
     
 
     public function importar_excel(request $request){
+        $permisos = $this->forpermisos('importar_bonos_nominas'); 
+            if($permisos=="importar_bonos_nominas")
+            {
+        
         $file=$request->file("urlxlsx");
        
 
@@ -197,18 +285,23 @@ class NominasController extends Controller
                 return back()->with("successExcel","¡Se guardaron los cambios correctamente!");
             }else{
                 return back()->with("warningExcel","¡Se guardaron los cambios correctamente!");
-            }
-
+            }            
+        }
+        }
+        else{
+            return redirect()->route('vernominas')->with("Errorpermisos","No se logro");   
         }
         
     }
 
     public function exportar_excel(request $request){
-
+        $permisos = $this->forpermisos('importar_bonos_nominas'); 
+            if($permisos=="importar_bonos_nominas")
+            {
         return Excel::download(new NominasExport($request->id), 'NominaExport.xlsx');
-
+        }
+        else{
+            return redirect()->route('vernominas')->with("Errorpermisos","No se logro");   
+        }
     }
-
-
-
 }
